backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet>=3.0.0
httpx==0.27.2
python-dotenv==1.0.1
python-dateutil==2.8.2
python-multipart==0.0.12
sentence-transformers==3.3.1
google-generativeai==0.3.2
numpy==1.26.4
pytest==8.3.5
pytest-asyncio==0.24.0
